+++ modules/features/onboarding/lib/src/domain/entity/photo/photographer.dart
@@ -1,0 +1,10 @@
+class Photographer {
+  final String id;
+  final String name;
+
+  Photographer({required this.id, required this.name});
+
+  factory Photographer.fromJson(Map<String, dynamic> json) {
+    return Photographer(id: json['id'], name: json['name']);
+  }
+}
