+++ modules/features/onboarding/lib/src/presentation/bloc/list_gallery_cubit.dart
@@ -1,5 +1,6 @@
-import 'package:entity/entity.dart';
+import 'package:entity_api/api.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:onboarding/src/domain/entity/entity.dart';
 import 'package:oni_api/oni_api.dart';
 
 class ListGalleryState {
@@ -23,75 +24,13 @@ class ListGalleryCubit extends Cubit<ListGalleryState> {
       },
     );
 
-    if (res is OniResultSuccess) {
+    if (res is ApiResultSuccess) {
       final photos = ListPhotoResponse.fromJson(res.data);
+      print('success');
+      print(photos.photos.first.id);
     } else {
       print('error');
       print(res);
     }
-  }
-}
-
-class ListPhotoResponse {
-  final int total;
-  final int totalPages;
-  final List<Photo> photos;
-
-  ListPhotoResponse({
-    required this.total,
-    required this.totalPages,
-    required this.photos,
-  });
-
-  factory ListPhotoResponse.fromJson(Map<String, dynamic> json) {
-    final results = json['results'] as List;
-    final List<Photo> photos = results
-        .map(
-          (item) => Photo.fromJson(item),
-        )
-        .toList();
-
-    return ListPhotoResponse(
-      total: json['total'],
-      totalPages: json['total_pages'],
-      photos: photos,
-    );
-  }
-}
-
-class Photo {
-  final String id;
-  final String slug;
-  final String description;
-  final int likes;
-  final Photographer photographer;
-
-  Photo({
-    required this.id,
-    required this.slug,
-    required this.description,
-    required this.likes,
-    required this.photographer,
-  });
-
-  factory Photo.fromJson(Map<String, dynamic> json) {
-    return Photo(
-      id: json['id'],
-      slug: json['slug'],
-      description: json['description'] ?? '',
-      likes: json['likes'],
-      photographer: Photographer.fromJson(json['user']),
-    );
-  }
-}
-
-class Photographer {
-  final String id;
-  final String name;
-
-  Photographer({required this.id, required this.name});
-
-  factory Photographer.fromJson(Map<String, dynamic> json) {
-    return Photographer(id: json['id'], name: json['name']);
   }
 }
