+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   oni_api:
     path: modules/shared/library/oni_api
   go_router: ^14.6.2
+  flutter_dotenv: ^5.0.2
 
 dev_dependencies:
   flutter_test:
@@ -59,7 +60,8 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
-
+  assets:
+    - .env
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
